Getting Composer back to the app root
Composer ought to lie at the root of the application for simplicity sake, as well as
in order to unleah its true power.

Compared to the previous branch state, "vendor" folder's been taken out of versionning
while "composer.lock" file's been taken in, as prescribed for Composer good practices.

As there are no single entry point to the application at this point, the web servers
settings have to be edited so that the application root path is provided to the PHP
include path.